Add role-based navigation helper
